refactor(dashboard): split the host/rig/node status sections out of views.py (#1105 Phase 3 D3, develop-v2)
Moves the system, worker, energy, xmrig-proxy, Tari and sync sections of the /api/state
payload — one contiguous 481-line run, 18 top-level names — into web/infra_views.py, and
their test classes 1:1 into tests/web/test_infra_views.py.

Boundary derived from the dependency graph, not the cut map's line ranges: the forward
closure of the map's D3 seed set is exactly those 18 names in ONE contiguous block with
zero non-member names interleaved. `_shorten` is also used by `build_pool_network`, which
stays, so views.py imports it back through the facade — the same shape D2 used for
`_LOW_HR_TITLE`, and no cycle, since infra_views.py imports nothing from views.py.

views.py stays the facade: build_state assembles these sections, and web/worker_detail.py's
existing import of `_rigforge_display`/`rigforge_update_for` keeps resolving through the
two noqa re-exports, so no consumer changes.

The moved source block is byte-identical. The moved test blocks are byte-identical modulo
three declared module-alias edits in TestBuildEnergy, which follow `views.config` to
`infra_views.config` — cosmetic, since both names are the same module object, but a test in
test_infra_views.py should not reach its target through views.

The six shared fixture builders are duplicated per the #1105 D2-D5 ruling, disclosed in the
new module's docstring, with #1459 as the tracked exit.
